Change default navigation to 2.12 only support (#339)
* Change default navigation to 2.12 only support

* Removed old navigation from tests

* Updated namespace toggle logic while GitRepo change

* Updated namespace/GitRepo nav logic to failed tests

* Changed dashboard navigation to actual GitRepo page

* Update Bundles navigation and App Bundles navigation

* Fixed failed tests with GitRepo Create navigation

* Increasing viewport width to get Delete option visible

* Fix P1_2 failed test

* Update locator for GitRepo path

* Skipping clusterGroup tests as per new changes in 2.12

* Fix label and title locator for failed tests

* Fix locators and Deployment status

* Added 2.12, head, alpha version filter to if-else block

* Fix the new section locator under secrets

* Fix Secret page filter search box

* Fix failed tests by adding wait and skipped tests which fails

* Fixed test before each part for SSH key creation

* Fix remove label and filter search box for secrets (#342)

* Fix 2.12 tests (#345)

* Fix test Fleet-51

* Fixed failed tests and updated logic for configmap when delete button is not present

* Fixed the login test


* Added head/2.12 support for test to run on 2.12

---------


* Fixed failed 2.12 tests (#351)

* Fixed failed 2.12 tests


* Added small wait after filter text

---------


* prime optimus alpha version to 2.12 branch (#355)

* Added backward Support along with 2.12


* Fixed 2.11 failed tests

* Fixed failed test by chaning namespace toggle place

* Fixed login test and other tests


* Updated timeout for resources to take some amount of time

* Fixed last failed test 51

---------

+++ tests/cypress/e2e/unit_tests/p0_fleet.spec.ts
@@ -19,6 +19,7 @@ export const appName = "nginx-keep";
 export const clusterName = "imported-0";
 export const branch = "main";
 export const path  = "nginx"
+export const sshString = ["Public key and private key for SSH", "Public key and private key for SSH authentication"]
 
 beforeEach(() => {
   cy.login();
@@ -36,8 +37,7 @@ describe('Test Fleet deployment on PUBLIC repos',  { tags: '@p0' }, () => {
       const path = "simple"
       const repoUrl = "https://github.com/rancher/fleet-examples"
 
-      cy.fleetNamespaceToggle('fleet-local');
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       // Adding check validate "Edit as Yaml" works
       cy.clickButton('Edit as YAML');
       cy.contains('apiVersion: fleet.cattle.io/v1alpha1').should('be.visible');
@@ -159,9 +159,10 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
       // Creating both known host keys in one loop
       secretKnownHostsKeys.forEach((secretKnownHostsKeys) => {
         cy.clickButton('Create');
-        cy.contains('Public key and private key for SSH').should('be.visible').click();
+        cy.contains(new RegExp(sshString.join('|'))).should('be.visible').click();
         cy.clickButton('Edit as YAML');
         cy.addYamlFile(secretKnownHostsKeys);
+        cy.wait(500);
         cy.clickButton('Create');
       });
     });
@@ -174,11 +175,10 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
         const repoName = 'local-cluster-fleet-141';
         const gitAuthType = 'ssh-key-knownhost';
     
-        cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+        cy.continuousDeliveryMenuSelection();
         
         // Create private repo using known host
-        cy.fleetNamespaceToggle('fleet-local');
-        cy.addFleetGitRepo({ repoName, repoUrl, gitAuthType, branch, path });
+        cy.addFleetGitRepo({ repoName, repoUrl, gitAuthType, branch, path, local: true });
         cy.clickButton('Create');
         cy.verifyTableRow(0, 'Active', '1/1');
         cy.checkGitRepoStatus(repoName, '1 / 1');
@@ -194,8 +194,7 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
           const gitAuthType = 'ssh-key-knownhost-missmatch';
 
           // Create private repo using known host
-          cy.fleetNamespaceToggle('fleet-local');
-          cy.addFleetGitRepo({ repoName, repoUrl, gitAuthType, branch, path });
+          cy.addFleetGitRepo({ repoName, repoUrl, gitAuthType, branch, path, local: true });
           cy.clickButton('Create');
 
           // Enrure that apps cannot be installed && error appears
@@ -234,9 +233,8 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
           cy.deleteAll(false);
           
           // Verify gitrepo is added using default knownhost
-          cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
-          cy.fleetNamespaceToggle('fleet-local');
-          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitAuthType, userOrPublicKey, pwdOrPrivateKey });
+          cy.continuousDeliveryMenuSelection();
+          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitAuthType, userOrPublicKey, pwdOrPrivateKey, local: true });
           cy.clickButton('Create');
           cy.checkGitRepoStatus(repoName, '1 / 1');
       })
@@ -251,9 +249,8 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
           
           // Verify gitrepo is canot be added when default knownhost exists
           // since it does not have ssh access
-          cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
-          cy.fleetNamespaceToggle('fleet-local');
-          cy.addFleetGitRepo({ repoName, repoUrl, branch, path});
+          cy.continuousDeliveryMenuSelection();
+          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true});
           cy.clickButton('Create');
           cy.verifyTableRow(0, /Error|Git Updating/, '0/0');
       })
@@ -271,8 +268,19 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
           cy.accesMenuSelection('local', 'Storage', 'Secrets');
           cy.nameSpaceMenuToggle('All Namespaces');
           cy.filterInSearchBox('ssh-key');
-          cy.deleteAll(false);
-          
+          cy.wait(500);
+
+          // Delete Secrets key if present
+          cy.get('body').then(($body) => {
+            const button = $body.find('[data-testid="sortable-table-promptRemove"]');
+            if (button.length > 0) {
+              cy.wrap(button).should('be.visible').then(() => {
+                cy.deleteAll(false);
+              });
+            } else {
+              cy.log("No Secrets are available for Delete.");
+            }
+          })
           // Delete default custom known-hosts
           cy.accesMenuSelection('local', 'Storage', 'ConfigMaps');
           cy.nameSpaceMenuToggle('All Namespaces');
@@ -290,9 +298,8 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
           cy.wait(500); // Needs time for previous command to finish
                 
           // Verify gitrepo is added using default knownhost
-          cy.accesMenuSelection('Continuous Delivery', 'Git Repos')
-          cy.fleetNamespaceToggle('fleet-local')
-          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitAuthType, userOrPublicKey, pwdOrPrivateKey })
+          cy.continuousDeliveryMenuSelection();
+          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitAuthType, userOrPublicKey, pwdOrPrivateKey, local: true })
           cy.clickButton('Create')
 
           // Ensure that apps cannot be installed && error appears
@@ -336,9 +343,8 @@ describe('Test gitrepos with cabundle', { tags: '@p0' }, () => {
       const path = 'simple'
       const tlsOption = "Specify additional certificates to be accepted"
       const tlsCertificate = "assets/cabundle-file.pem"
-  
-      cy.fleetNamespaceToggle('fleet-local');
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, tlsOption, tlsCertificate });
+
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, tlsOption, tlsCertificate, local: true });
       cy.clickButton('Create');
       cy.verifyTableRow(0, 'Active', '1/1');
       cy.accesMenuSelection('local', 'Storage', 'Secrets');
@@ -362,9 +368,8 @@ describe('Test gitrepos with cabundle', { tags: '@p0' }, () => {
       const repoUrl = 'https://github.com/rancher/fleet-examples'
       const branch = 'master'
       const path = 'simple'
-  
-      cy.fleetNamespaceToggle('fleet-local');
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       cy.clickButton('Create');
       cy.verifyTableRow(0, 'Active', '1/1');
       cy.accesMenuSelection('local', 'Storage', 'Secrets');
@@ -431,9 +436,9 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
         });
 
         // Gitrepo creation via YAML
-        cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+        cy.continuousDeliveryMenuSelection();
         cy.fleetNamespaceToggle('fleet-local');
-        cy.clickButton('Add Repository');
+        cy.clickCreateGitRepo();
         cy.clickButton('Edit as YAML');
         cy.addYamlFile('assets/webhook-tests/webhook_test_disable_polling.yaml');
         cy.clickButton('Create');
@@ -481,9 +486,9 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
         kubectl delete secrets -n cattle-fleet-system gitjob-webhook{enter}');
 
         // Gitrepo creation via YAML
-        cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+        cy.continuousDeliveryMenuSelection();
         cy.fleetNamespaceToggle('fleet-local');
-        cy.clickButton('Add Repository');
+        cy.clickCreateGitRepo();
         cy.clickButton('Edit as YAML');
         
         cy.addYamlFile('assets/webhook-tests/webhook_test_webhook_secret_in_repo.yaml');
@@ -537,9 +542,9 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
         kubectl create secret generic gitjob-webhook -n cattle-fleet-system --from-literal=github=wrong-webhook-secret{enter}');
 
         // Gitrepo creation via YAML
-        cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+        cy.continuousDeliveryMenuSelection();
         cy.fleetNamespaceToggle('fleet-local');
-        cy.clickButton('Add Repository');
+        cy.clickCreateGitRepo();
         cy.clickButton('Edit as YAML');
         
         cy.addYamlFile('assets/webhook-tests/webhook_test_disable_polling.yaml');
@@ -586,9 +591,8 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
       it('Fleet-145: Test Fleet job cleanup', { tags: '@fleet-145' }, () => {
 
         const repoName = 'local-145-test-job-cleanup';
-  
-        cy.fleetNamespaceToggle('fleet-local');
-        cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+
+        cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
         cy.clickButton('Create');
         cy.verifyTableRow(0, 'Active', '1/1');
 
@@ -603,8 +607,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
 
       const repoName = 'local-146-test-job-cleanup';
 
-      cy.fleetNamespaceToggle('fleet-local');
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       cy.clickButton('Create');
       cy.verifyTableRow(0, 'Active', '1/1');
 
@@ -632,7 +635,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
 
       // Gitrepo adddition via YAML
       cy.fleetNamespaceToggle('fleet-local');
-      cy.clickButton('Add Repository');
+      cy.clickCreateGitRepo();
       cy.clickButton('Edit as YAML');
       cy.addYamlFile('assets/disable_polling.yaml');
       cy.clickButton('Create');
@@ -645,7 +648,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
       cy.exec('bash assets/disable_polling_setting_5_replicas.sh').then((result) => {
         cy.log(result.stdout, result.stderr);
       });
-      cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+      cy.continuousDeliveryMenuSelection();
       cy.fleetNamespaceToggle('fleet-local');
       cy.checkGitRepoStatus(repoName, '1 / 1', '1 / 1');
 
@@ -659,9 +662,8 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
   
       const repoName = 'local-148-test-unsuscessful-job-is-not-deleted';
       const path = 'qa-test-apps/nginx-app-bad-path';
-  
-      cy.fleetNamespaceToggle('fleet-local');
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       cy.clickButton('Create');
       cy.verifyTableRow(0, /Git Updating|Error/, '0/0');
       
@@ -737,8 +739,7 @@ if (!/\/2\.8/.test(Cypress.env('rancher_version'))) {
         const repoName = 'simple-chart-secret';
   
         // Deploy Gitrepo
-        cy.fleetNamespaceToggle('fleet-local');
-        cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+        cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
         cy.clickButton('Create');
         cy.verifyTableRow(0, 'Active', '1/1');
