+++ tests/cypress/e2e/unit_tests/p1_fleet.spec.ts
@@ -32,10 +32,9 @@ beforeEach(() => {
 describe('Test resource behavior after deleting GitRepo using keepResources option using YAML', { tags: '@p1' }, () => {
   qase(68, 
     it('Fleet-68: Test RESOURCES will be KEPT and NOT be DELETED after GitRepo is deleted when keepResources: true used in the GitRepo yaml file.', { tags: '@fleet-68' }, () => {
-      cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+      cy.continuousDeliveryMenuSelection();
       cy.fleetNamespaceToggle('fleet-local')
-      cy.clickButton('Add Repository');
-      cy.contains('Git Repo:').should('be.visible');
+      cy.clickCreateGitRepo();
       cy.clickButton('Edit as YAML');
       cy.addYamlFile('assets/git-repo-keep-resources-true.yaml');
       cy.clickButton('Create');
@@ -70,8 +69,7 @@ describe('Test resource behavior after deleting GitRepo using keepResources opti
       qase(qase_id,
         it(`Fleet-${qase_id}: Test ${test_explanation}`, { tags: `@fleet-${qase_id}` }, () => {
           const repoName = `local-cluster-fleet-${qase_id}`
-          cy.fleetNamespaceToggle('fleet-local')
-          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, keepResources });
+          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, keepResources, local: true });
           cy.clickButton('Create');
           cy.checkGitRepoStatus(repoName, '1 / 1', '1 / 1');
           cy.checkApplicationStatus(appName);
@@ -102,8 +100,7 @@ describe('Test Self-Healing of resource modification when correctDrift option us
       qase(qase_id,
         it(`Fleet-${qase_id}: Test ${test_explanation}`, { tags: `@fleet-${qase_id}` }, () => {
           const repoName = `local-cluster-correct-${qase_id}`
-          cy.fleetNamespaceToggle('fleet-local')
-          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, correctDrift });
+          cy.addFleetGitRepo({ repoName, repoUrl, branch, path, correctDrift, local: true });
           cy.clickButton('Create');
           cy.checkGitRepoStatus(repoName, '1 / 1', '1 / 1');
           cy.checkApplicationStatus(appName);
@@ -129,8 +126,7 @@ describe('Test Self-Healing of resource modification when correctDrift option us
   qase(77,
     it("Fleet-77: Test MODIFICATION to resources will be self-healed when correctDrift is set to true in existing GitRepo.", { tags: '@fleet-77', retries: 1 }, () => {
       const repoName = "local-cluster-correct-77"
-      cy.fleetNamespaceToggle('fleet-local')
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       cy.clickButton('Create');
       cy.checkGitRepoStatus(repoName, '1 / 1', '1 / 1');
       cy.checkApplicationStatus(appName);
@@ -169,8 +165,7 @@ describe('Test resource behavior after deleting GitRepo using keepResources opti
   qase(71,
     it("Fleet-71: Test RESOURCES will be KEPT and NOT be DELETED after GitRepo is deleted when keepResources is set to true in existing GitRepo.", { tags: '@fleet-71' }, () => {
       const repoName = "local-cluster-keep-71"
-      cy.fleetNamespaceToggle('fleet-local')
-      cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+      cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
       cy.clickButton('Create');
       cy.checkGitRepoStatus(repoName, '1 / 1', '1 / 1');
       cy.checkApplicationStatus(appName);
@@ -213,14 +208,13 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version'))) {
         const branch = 'master'
         const path = 'imagescans'
 
-        cy.fleetNamespaceToggle('fleet-local');
-        cy.addFleetGitRepo({ repoName, repoUrl, branch, path });
+        cy.addFleetGitRepo({ repoName, repoUrl, branch, path, local: true });
         cy.clickButton('Create');
         cy.verifyTableRow(0, 'Error', '1/1');
         cy.accesMenuSelection('local', 'Workloads');
         cy.nameSpaceMenuToggle('All Namespaces');
         cy.filterInSearchBox('fleet-controller');
-        cy.verifyTableRow(0, 'Running', 'fleet-controller')
+        cy.verifyTableRow(0, /Running|Active/, 'fleet-controller')
         cy.deleteAllFleetRepos();
       })
     )
@@ -270,8 +264,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
           it(`Fleet-${qase_id}: Test private helm registries for \"helmRepoURLRegex\" matches with \"${test_explanation}\" URL specified in fleet.yaml file`, { tags: `@fleet-${qase_id}` }, () => {;
             // Positive test using matching regex
             helmUrlRegex = helmUrlRegex_matching
-            cy.fleetNamespaceToggle('fleet-local');
-            cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitOrHelmAuth, gitAuthType, userOrPublicKey, pwdOrPrivateKey, helmUrlRegex });
+            cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitOrHelmAuth, gitAuthType, userOrPublicKey, pwdOrPrivateKey, helmUrlRegex, local: true });
             cy.clickButton('Create');
             cy.verifyTableRow(0, 'Active', /([1-9]\d*)\/\1/);
             cy.accesMenuSelection('local', 'Storage', 'ConfigMaps');
@@ -283,8 +276,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
             cy.deleteAllFleetRepos();
             // Negative test using non-matching regex 1234.*
             helmUrlRegex = '1234.*'
-            cy.fleetNamespaceToggle('fleet-local');
-            cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitOrHelmAuth, gitAuthType, userOrPublicKey, pwdOrPrivateKey, helmUrlRegex });
+            cy.addFleetGitRepo({ repoName, repoUrl, branch, path, gitOrHelmAuth, gitAuthType, userOrPublicKey, pwdOrPrivateKey, helmUrlRegex, local: true });
             cy.clickButton('Create');
             cy.get('.text-error', { timeout: 120000 }).should('contain', 'code: 401');
             cy.deleteAllFleetRepos();
@@ -402,7 +394,7 @@ if (!/\/2\.9/.test(Cypress.env('rancher_version'))) {
 
             // Adding more wait for 30seconds to capture the error if occurred after modifying the resources.
             cy.wait(30000);
-            cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+            cy.continuousDeliveryMenuSelection();
             cy.verifyTableRow(0, 'Active', repoName);
 
             // Check All clusters are in healthy state after performing any modification to the resources.
@@ -425,10 +417,25 @@ if (!/\/2\.9/.test(Cypress.env('rancher_version'))) {
             dsAllClusterList.forEach((dsCluster) => {
               // Adding wait to load page correctly to avoid interference with hamburger-menu.
               cy.wait(500);
-              cy.accesMenuSelection(dsCluster, "Service Discovery", "Services");
+              cy.accesMenuSelection(dsCluster, resourceLocation, resourceType);
               cy.nameSpaceMenuToggle(resourceNamespace);
               cy.filterInSearchBox(resourceName);
-              cy.deleteAll(false);
+              cy.wait(500);
+              if (resourceType === 'ConfigMaps') {
+                cy.deleteConfigMap(resourceName);
+              }
+              else {
+                cy.get('body').then(($body) => {
+                  const button = $body.find('[data-testid="sortable-table-promptRemove"]');
+                  if (button.length > 0) {
+                    cy.wrap(button).should('be.visible').then(() => {
+                      cy.deleteAll(false);
+                    });
+                  } else {
+                    cy.log("No Service(s) available for Delete.");
+                  }
+                })
+              }
               cy.clickNavMenu(['Cluster', 'Projects/Namespaces']);
               cy.filterInSearchBox(resourceNamespace);
               cy.deleteAll(false);
@@ -458,7 +465,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
       });
 
       cy.fleetNamespaceToggle('fleet-local');
-      cy.clickButton('Add Repository');
+      cy.clickCreateGitRepo();
 
       // Pass YAML file (no previous additions of Name, urls or paths)
       cy.clickButton('Edit as YAML');
@@ -497,7 +504,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
           cy.screenshot('Screenshot AFTER reloading should be 2/2');
           cy.verifyTableRow(0, 'Active', '2/2');
 
-          cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+          cy.continuousDeliveryMenuSelection();
           cy.fleetNamespaceToggle('fleet-local');
           cy.open3dotsMenu('test-disable-polling', 'Force Update');
           cy.wait(2000); // Wait to let time for Update to take effect.
@@ -538,7 +545,7 @@ if (!/\/2\.7/.test(Cypress.env('rancher_version')) && !/\/2\.8/.test(Cypress.env
 
           prepareGithubRepoReplicas()
 
-          cy.accesMenuSelection('Continuous Delivery', 'Git Repos');
+          cy.continuousDeliveryMenuSelection();
           cy.fleetNamespaceToggle('fleet-local');
           cy.open3dotsMenu(repoName, 'Pause');
